questions/views.py
from rest_framework import generics
from .serializers import QuestionSerializer
from .models import Question

# Create your views here.

class QuestionList(generics.ListAPIView):
    queryset = Question.objects.all()
    serializer_class = QuestionSerializer

class RetrieveQuestion(generics.RetrieveUpdateDestroyAPIView):
    queryset = Question.objects.all()
    serializer_class = QuestionSerializer
    