requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.8
django-cleanup==5.2.0
django-filter==21.1
djangorestframework==3.12.4
drf-spectacular==0.20.1
drf-spectacular-sidecar==2021.9.27
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.1.0
Markdown==3.3.4
Pillow==8.3.2
psycopg2-binary==2.9.1
pyrsistent==0.18.0
pytz==2021.3
PyYAML==5.4.1
sqlparse==0.4.2
uritemplate==4.0.0
whitenoise==5.3.0
